Moved the units tests to the module tests/__init__.py
(cherry picked from commit ebdc4fab694b76cee0b93b9279288112ed30ea4e)
